Fix parsing of python f-strings (#160)
* Fix parsing of python f-strings

Double-quoted f-strings mostly worked, but single-quoted did not:

    a = f'what'

Reported that `f'what` was mispelled. I think this is due to:

1. the `(strings) @string` query give back in the whole node string
   node - including `f'`
2. the word-extraction in parser.rs (`processor.extract_words()`)
   does not consider `'` a word separator (rightly so)

(2) explains why double-quote f-strings works - the word splitting
split on `"`, resulting in "f" and "what", and I assume single letter
words is not checked.

(1) can be confirmed in the tree-sitter playgound [1], where we
also see that there is an `string_content` node type more suitable.

Using the `string_content` node means that code inside interpolation
units wont be spell-checked, but this aligns with the current behavior
that only definitions, comments, and strings are checked.

[1] https://tree-sitter.github.io/tree-sitter/7-playground.html

* Unit test for python f-string spell checking